test(fuzz): checksum validation errors happen
In the current fuzzing setup, it is expected that checksum validation
errors will happen. This is perhaps obvious, but worth stating: this
library is functioning correctly when checksum errors are detected.
(To be clear, the current fuzzing tests are not setup in such a way
to assess if/when checksum errors should be detected.)